Upgrade to tree-sitter 0.26, drop stack-graphs, add 5 new languages
BREAKING: Removes tree-sitter-stack-graphs dependency entirely.
The cross-file resolver + type propagation handles resolution with
better performance and comparable quality.

tree-sitter upgrade: 0.24 -> 0.26
- Supports ABI 13-15 grammars (was limited to 13-14)
- All existing grammars updated to latest compatible versions

New languages (5): Lua, PHP, Dart, Scala, HCL
- Total supported: 18 languages
- All use the same extractor pattern (walker + scope stack)

Performance impact (no more stack-graphs timeout):
- rust-bat: 60s -> 0.5s (110x faster)
- js-lodash: 60s -> 0.35s (172x faster)
- ts-typeorm: 60s -> 1.3s (47x faster)
- python-httpie: 3.8s -> 0.09s (40x faster)
- js-express: 10.4s -> 0.05s (192x faster)

Edge quality comparison (cross-file edges preserved):
- All projects maintain or improve cross-file edge counts
- The cross-file resolver + type propagation already handled
  95%+ of what stack-graphs found

New project benchmarks:
- php-laravel: 28,840 callables, 767ms
- scala-spark: 98,722 callables, 59s (8831 files)
- hcl-terraform-aws: 104,073 callables, 79s (12194 files)

177 tests passing. --stack-graphs flag retained (auto/on/off) but
all modes now no-op since the resolver is removed.

+++ crates/cgg-lang/src/plugins/bash.rs
@@ -100,7 +100,7 @@ impl<'a> BashWalker<'a> {
                 .or_else(|| {
                     let count = node.child_count();
                     for i in 0..count {
-                        let c = node.child(i).unwrap();
+                        let c = node.child(i as u32).unwrap();
                         if c.kind() == "word" || c.kind() == "string" {
                             return Some(c);
                         }
